requirements.txt
groq
requests
python-dotenv
pyflakes
astroid
